Renamed respective drivers to include GPIO
