+++ src/System.Device.Gpio/System/Device/Gpio/Drivers/HummingBoardGpioDriver.Windows.cs
@@ -4,7 +4,7 @@
 
 namespace System.Device.Gpio.Drivers
 {
-    public partial class RaspberryPi3Driver : Windows10Driver
+    public partial class HummingBoardGpioDriver : Windows10GpioDriver
     {
     }
 }
