+++ src/System.Device.Gpio/System/Device/Gpio/Drivers/RaspberryPi3GpioDriver.Windows.cs
@@ -4,7 +4,7 @@
 
 namespace System.Device.Gpio.Drivers
 {
-    public partial class HummingBoardDriver : Windows10Driver
+    public partial class RaspberryPi3GpioDriver : Windows10GpioDriver
     {
     }
 }
