+++ src/System.Device.Gpio/System/Device/Gpio/Drivers/Windows10GpioDriver.Linux.cs
@@ -3,23 +3,17 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Threading;
-using System.Threading.Tasks;
 
 namespace System.Device.Gpio.Drivers
 {
-    public class UnixDriver : GpioDriver
+    public class Windows10GpioDriver : GpioDriver
     {
-        public UnixDriver()
+        public Windows10GpioDriver()
         {
             throw new PlatformNotSupportedException();
         }
 
         protected internal override int PinCount => throw new PlatformNotSupportedException();
-
-        protected override void Dispose(bool disposing)
-        {
-            throw new PlatformNotSupportedException();
-        }
 
         protected internal override void AddCallbackForPinValueChangedEvent(int pinNumber, PinEventTypes eventType, PinChangeEventHandler callback)
         {
@@ -67,11 +61,6 @@ namespace System.Device.Gpio.Drivers
         }
 
         protected internal override WaitForEventResult WaitForEvent(int pinNumber, PinEventTypes eventType, CancellationToken cancellationToken)
-        {
-            throw new PlatformNotSupportedException();
-        }
-
-        protected internal override ValueTask<WaitForEventResult> WaitForEventAsync(int pinNumber, PinEventTypes eventType, CancellationToken cancellationToken)
         {
             throw new PlatformNotSupportedException();
         }
